feat(TDP-5398): move UserDataSetMetadata converter to backend-service
* allows UserDataSetMetadata conversion in DataSetAPI

* rename DataSetConversions to UserMetadataConversion
